refactor(connector): put classes definition in types.ts and upgrade silverhand/essentials dependency
