+++ packages/connector-schemas/package.json
@@ -23,7 +23,7 @@
   },
   "devDependencies": {
     "@silverhand/eslint-config": "1.0.0-rc.2",
-    "@silverhand/essentials": "^1.1.6",
+    "@silverhand/essentials": "^1.2.0",
     "@silverhand/ts-config": "1.0.0-rc.2",
     "eslint": "^8.21.0",
     "lint-staged": "^13.0.0",
