+++ packages/connector-schemas/src/index.ts
@@ -1,3 +1,2 @@
 export * from './types';
 export * from './error';
-export * from './class';
